ACTIN-68: Add ActionableFunctionsTest
